+++ src/main/java/com/github/jlangch/venice/impl/functions/IOFunctions.java
@@ -691,7 +691,7 @@ public class IOFunctions {
                         "(io/file-matches-glob? \"**/*.log\" \"x/y/file.log\")",
                         "(io/file-matches-glob? (io/glob-path-matcher\"*.log\") (io/file \"file.log\"))",
                         "(io/file-matches-glob? (io/glob-path-matcher\"**/*.log\") (io/file \"x/y/file.log\"))")
-                   .seeAlso("io/glob-path-matcher", "io/list-files-glob")
+                    .seeAlso("io/glob-path-matcher", "io/list-files-glob")
                     .build()
         ) {
             @Override